Angle calculator fully working, just need to add check for 180.

+++ PolygonChecker/triangleSolver.h
@@ -1,3 +1,3 @@
 #pragma once
 char* analyzeTriangle(int side1, int side2, int side3);
-int  angleCalculatorTriangle(int* trianglesides);
+double  angleCalculatorTriangle(int* trianglesides);
